faiss_service/requirements.txt
fastapi==0.112.0
uvicorn==0.30.0
faiss-cpu==1.8.0
numpy==1.26.4